requirements.txt
Eve==1.1
requests==2.22.0
huey==2.2.0
gunicorn==20.0.4
py-bcrypt==0.4
flask-cors==3.0.8